extracted list into separate component
