+++ src/components/pokedex-display/style.ts
@@ -1,0 +1,5 @@
+export const searchBarStyle = {
+  height: "65px",
+  margin: "20px 0",
+  boxSizing: "border-box",
+};
