+++ src/pages/home-page/style.ts
@@ -2,14 +2,8 @@ import { SxProps, Theme } from "@mui/material";
 
 export const homePageContainerStyle: SxProps<Theme> = {
   display: "flex",
-  width: "100vw",
+  width: "100%",
   boxSizing: "border-box",
   p: "0 130px",
   justifyContent: "center",
-};
-
-export const searchBarStyle = {
-  height: "65px",
-  margin: "20px 0",
-  boxSizing: "border-box",
 };
